書き方のルール5を適用
