+++ src/hooks/useCreateDeleteMe.js
@@ -1,0 +1,13 @@
+const useCreateDeleteMe = (subjectAndPointList, setSubjectAndPointList) => {
+  const createDeleteMe = (index) => {
+    return () => {
+      const _subjectAndPointList = [...subjectAndPointList];
+      _subjectAndPointList.splice(index, 1);
+      setSubjectAndPointList(_subjectAndPointList);
+    }
+  }
+
+  return createDeleteMe;
+}
+
+export default useCreateDeleteMe;
